requirements.txt
# 台灣股市每日自動抓取 - Python 依賴套件
# 安裝方式: pip install -r requirements.txt

# Google Sheets API
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0

# HTTP 請求
requests>=2.31
